moved login code to its own component - Login
